requirements.txt
Flask==3.1.0
Flask-Compress==1.17
requests==2.32.3
PyYAML==6.0.2
